# Conflicts:
#	src/main/resources/application.yaml

+++ src/main/java/uk/gov/hmcts/reform/hmc/controllers/HearingManagementController.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.reform.hmc.controllers;
 
+import com.microsoft.applicationinsights.core.dependencies.google.common.collect.Lists;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
@@ -18,79 +19,92 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import uk.gov.hmcts.reform.hmc.domain.model.enums.HearingStatus;
+import uk.gov.hmcts.reform.hmc.exceptions.ValidationError;
 import uk.gov.hmcts.reform.hmc.model.DeleteHearingRequest;
 import uk.gov.hmcts.reform.hmc.model.GetHearingResponse;
 import uk.gov.hmcts.reform.hmc.model.GetHearingsResponse;
 import uk.gov.hmcts.reform.hmc.model.HearingRequest;
 import uk.gov.hmcts.reform.hmc.model.HearingResponse;
 import uk.gov.hmcts.reform.hmc.model.UpdateHearingRequest;
+import uk.gov.hmcts.reform.hmc.service.AccessControlService;
 import uk.gov.hmcts.reform.hmc.service.HearingManagementService;
 
+import java.util.List;
 import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Size;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static uk.gov.hmcts.reform.hmc.constants.Constants.AMEND_HEARING;
-import static uk.gov.hmcts.reform.hmc.constants.Constants.DELETE_HEARING;
-import static uk.gov.hmcts.reform.hmc.constants.Constants.REQUEST_HEARING;
-import static uk.gov.hmcts.reform.hmc.exceptions.ValidationError.CASE_REF_EMPTY;
-import static uk.gov.hmcts.reform.hmc.exceptions.ValidationError.CASE_REF_INVALID;
-import static uk.gov.hmcts.reform.hmc.exceptions.ValidationError.CASE_REF_INVALID_LENGTH;
+import static uk.gov.hmcts.reform.hmc.service.AccessControlServiceImpl.HEARING_MANAGER;
+import static uk.gov.hmcts.reform.hmc.service.AccessControlServiceImpl.HEARING_VIEWER;
+import static uk.gov.hmcts.reform.hmc.service.AccessControlServiceImpl.LISTED_HEARING_VIEWER;
 
 @RestController
 @Validated
 public class HearingManagementController {
 
     private final HearingManagementService hearingManagementService;
+    private final AccessControlService accessControlService;
 
-    public HearingManagementController(HearingManagementService hearingManagementService) {
+    public HearingManagementController(HearingManagementService hearingManagementService,
+                                       AccessControlService accessControlService) {
         this.hearingManagementService = hearingManagementService;
+        this.accessControlService = accessControlService;
     }
 
     @GetMapping(path = "/hearing/{id}", produces = APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @ApiResponses(value = {
         @ApiResponse(code = 204, message = "Hearing id is valid"),
-        @ApiResponse(code = 404, message = "Hearing id not found"),
-        @ApiResponse(code = 400, message = "Invalid hearing id")
+        @ApiResponse(code = 404, message = ValidationError.HEARING_ID_NOT_FOUND),
+        @ApiResponse(code = 400, message = ValidationError.INVALID_HEARING_ID_DETAILS)
     })
     public ResponseEntity<GetHearingResponse> getHearing(@PathVariable("id") Long hearingId,
                                                          @RequestParam(value = "isValid",
                                                              defaultValue = "false") boolean isValid) {
+        if (!isValid) {
+            // Only verify access if the user is requesting more than just confirmation of a valid hearing id
+            String status = hearingManagementService.getStatus(hearingId);
+            List<String> requiredRoles = Lists.newArrayList(HEARING_VIEWER);
+            if (HearingStatus.LISTED.name().equals(status)) {
+                requiredRoles.add(LISTED_HEARING_VIEWER);
+            }
+
+            accessControlService.verifyHearingCaseAccess(hearingId, requiredRoles);
+        }
         return hearingManagementService.getHearingRequest(hearingId, isValid);
     }
 
     @PostMapping(path = "/hearing", consumes = APPLICATION_JSON_VALUE, produces = APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.CREATED)
     @ApiResponses(value = {
-        @ApiResponse(code = 201, message = "Hearing Id is created"),
-        @ApiResponse(code = 400, message = "Invalid hearing details found")
+        @ApiResponse(code = 201, message = "Hearing successfully created"),
+        @ApiResponse(code = 400,
+            message = "One or more of the following reasons:"
+                + "\n1) " + ValidationError.INVALID_HEARING_REQUEST_DETAILS
+                + "\n2) " + ValidationError.INVALID_HEARING_WINDOW
+                + "\n3) " + ValidationError.INVALID_ORG_INDIVIDUAL_DETAILS
+        )
     })
     public HearingResponse saveHearing(@RequestBody @Valid HearingRequest createHearingRequest) {
-        hearingManagementService.verifyAccess(getCaseRef(createHearingRequest));
-        HearingResponse hearingResponse = hearingManagementService.saveHearingRequest(createHearingRequest);
-        hearingManagementService.sendRequestToHmiAndQueue(hearingResponse.getHearingRequestId(), createHearingRequest,
-                REQUEST_HEARING
-        );
-        return hearingResponse;
+        accessControlService.verifyCaseAccess(getCaseRef(createHearingRequest), Lists.newArrayList(HEARING_MANAGER));
+        return hearingManagementService.saveHearingRequest(createHearingRequest);
     }
 
     @DeleteMapping(path = "/hearing/{id}", consumes = APPLICATION_JSON_VALUE, produces = APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.OK)
     @ApiResponses(value = {
         @ApiResponse(code = 200, message = "Hearing cancellation processed"),
-        @ApiResponse(code = 400, message = "Invalid hearing details found"),
-        @ApiResponse(code = 404, message = "Hearing id not found"),
-        @ApiResponse(code = 500, message = "Error occurred on the server")
+        @ApiResponse(code = 400, message = ValidationError.INVALID_HEARING_REQUEST_DETAILS),
+        @ApiResponse(code = 404, message = ValidationError.HEARING_ID_NOT_FOUND),
+        @ApiResponse(code = 500, message = ValidationError.INTERNAL_SERVER_ERROR)
     })
     public HearingResponse deleteHearing(@PathVariable("id") Long hearingId,
                                          @RequestBody @Valid DeleteHearingRequest deleteRequest) {
-        HearingResponse hearingResponse = hearingManagementService.deleteHearingRequest(
+        accessControlService.verifyHearingCaseAccess(hearingId, Lists.newArrayList(HEARING_MANAGER));
+        return hearingManagementService.deleteHearingRequest(
             hearingId, deleteRequest);
-        hearingManagementService.sendRequestToHmiAndQueue(deleteRequest, hearingId, DELETE_HEARING);
-
-        return hearingResponse;
     }
 
     /**
@@ -106,16 +120,29 @@ public class HearingManagementController {
     @ApiOperation(value = "Get hearings")
     @ApiResponses({
         @ApiResponse(code = 200, message = "Success (with content)"),
-        @ApiResponse(code = 400, message = "Invalid request")
+        @ApiResponse(code = 400,
+            message = "One or more of the following reasons:"
+                + "\n1) " + ValidationError.INVALID_HEARING_REQUEST_DETAILS
+                + "\n2) " + ValidationError.CASE_REF_EMPTY
+                + "\n3) " + ValidationError.CASE_REF_INVALID_LENGTH
+                + "\n4) " + ValidationError.CASE_REF_INVALID
+        )
     })
-    public GetHearingsResponse getHearings(@PathVariable("ccdCaseRef")
-                                           @Valid
-                                           @NotEmpty(message = CASE_REF_EMPTY)
-                                           @Size(min = 16, max = 16, message = CASE_REF_INVALID_LENGTH)
-                                           @LuhnCheck(message = CASE_REF_INVALID, ignoreNonDigitCharacters = false)
+    public GetHearingsResponse getHearings(@PathVariable("ccdCaseRef") @Valid
+                                           @NotEmpty(message = ValidationError.CASE_REF_EMPTY)
+                                           @Size(min = 16, max = 16, message = ValidationError.CASE_REF_INVALID_LENGTH)
+                                           @LuhnCheck(message = ValidationError.CASE_REF_INVALID,
+                                               ignoreNonDigitCharacters = false)
                                                String ccdCaseRef,
                                            @RequestParam(required = false)
                                                String status) {
+        if (HearingStatus.LISTED.name().equals(status)) {
+            accessControlService.verifyCaseAccess(ccdCaseRef, Lists.newArrayList(
+                HEARING_VIEWER,
+                LISTED_HEARING_VIEWER));
+        } else {
+            accessControlService.verifyCaseAccess(ccdCaseRef, Lists.newArrayList(HEARING_VIEWER));
+        }
         return hearingManagementService.getHearings(ccdCaseRef, status);
     }
 
@@ -123,30 +150,31 @@ public class HearingManagementController {
     @ResponseStatus(HttpStatus.CREATED)
     @ApiResponses(value = {
         @ApiResponse(code = 201, message = "Hearing successfully updated"),
-        @ApiResponse(code = 400, message = "Invalid hearing details found"),
-        @ApiResponse(code = 404, message = "Hearing id not found"),
-        @ApiResponse(code = 500, message = "Error occurred on the server")
+        @ApiResponse(code = 400, message = ValidationError.INVALID_HEARING_REQUEST_DETAILS),
+        @ApiResponse(code = 404, message = ValidationError.CASE_NOT_FOUND),
+        @ApiResponse(code = 500, message = ValidationError.INTERNAL_SERVER_ERROR)
     })
     public HearingResponse updateHearing(@RequestBody @Valid UpdateHearingRequest hearingRequest,
                                          @PathVariable("id") Long hearingId) {
-        HearingResponse hearingResponse = hearingManagementService.updateHearingRequest(hearingId, hearingRequest);
-        hearingManagementService.sendRequestToHmiAndQueue(hearingId, hearingRequest, AMEND_HEARING);
-        return hearingResponse;
+        accessControlService.verifyHearingCaseAccess(hearingId, Lists.newArrayList(HEARING_MANAGER));
+        return hearingManagementService.updateHearingRequest(hearingId, hearingRequest);
     }
 
     @PostMapping(path = "/hearingActualsCompletion/{id}")
     @ResponseStatus(HttpStatus.OK)
     @ApiResponses(value = {
         @ApiResponse(code = 200, message = "Success (with no content)"),
-        @ApiResponse(code = 404, message = "001 No such id"),
-        @ApiResponse(code = 400, message = "Invalid hearing details found"),
-        @ApiResponse(code = 400, message = "002 invalid status"),
-        @ApiResponse(code = 400, message = "003 missing hearing day actuals"),
-        @ApiResponse(code = 400, message = "004 unexpected hearing day actuals"),
-        @ApiResponse(code = 400, message = "005 missing hearing outcome"),
-        @ApiResponse(code = 500, message = "Error occurred on the server")
+        @ApiResponse(code = 404, message = ValidationError.HEARING_ACTUALS_ID_NOT_FOUND),
+        @ApiResponse(code = 400,
+            message = "One or more of the following reasons:"
+                + "\n1) " + ValidationError.INVALID_HEARING_REQUEST_DETAILS
+                + "\n2) " + ValidationError.HEARING_ACTUALS_INVALID_STATUS
+                + "\n3) " + ValidationError.HEARING_ACTUALS_UN_EXPECTED
+                + "\n4) " + ValidationError.HEARING_ACTUALS_MISSING_HEARING_OUTCOME),
+        @ApiResponse(code = 500, message = ValidationError.INTERNAL_SERVER_ERROR)
     })
     public ResponseEntity hearingCompletion(@PathVariable("id") Long hearingId) {
+        accessControlService.verifyHearingCaseAccess(hearingId, Lists.newArrayList(HEARING_MANAGER));
         return hearingManagementService.hearingCompletion(hearingId);
     }
